feat: send non-log files to image parsing service
This commit modifies the `put_line_log_to_db.py` file to send non-log files to the image parsing service. If the object key does not end with `.log`, the code will send an event to the FIFO SQS for image parsing. Additionally, if the message type is "image", it will send the message to the image parsing service. This change improves the functionality of the code by ensuring that non-log files are properly handled and sent for image parsing.