refactor: update logger configuration in code files
This commit refactors the code files in the `put_line_log_to_db_function` directory to update the logger configuration. The `logging.basicConfig` method is replaced with `logger.setLevel("INFO")` to set the log level to "INFO". This change improves the logging functionality by ensuring that only log messages with a level of "INFO" or higher are displayed.