refactor: add new sticker to end_recording in stickers.json
This commit adds a new sticker to the `end_recording` array in the `stickers.json` file. The sticker has a `packageId` of "11539" and a `stickerId` of "52114117". This modification enhances the functionality of the code by including the new sticker in the list of stickers that indicate the end of a recording.